+++ NEWS.md
@@ -14,10 +14,19 @@ It is now also necessary to explicitly enable the use of weights via the new hyp
 Note that this is a breaking change, but explicitly stating the intention to use the weights is the less error-prone solution in the long run.
 
 
+* feat: The `partition()` function creates training, test and validation sets.
 * refactor: Optimize runtime of fixing factor levels.
 * refactor: Optimize runtime of setting row roles.
 * refactor: Optimize runtime of marshalling.
 * refactor: Optimize runtime of `Task$col_info`
+* BREAKING CHANGE: the predict time of the learner now stores the cumulative duration for
+  all predict sets (#992).
+* feat: `$internal_valid_task` can now be set to an `integer` vector.
+* deprecated the `$divide()` method
+* fix: `Task$cbind()` now works with non-standard primary keys
+  for `data.frames` (#961).
+* fix: Triggering of fallback learner now has log-level "info"
+  instead of "debug" (#972)
 
 # mlr3 0.20.2